moved registration code to one function in mainRegistration
